+++ lib/l10n/app_localizations_de.dart
@@ -3797,13 +3797,13 @@ class AppLocalizationsDe extends AppLocalizations {
   String get reefDevicesKindReefFactory => 'ReefFactory';
 
   @override
-  String get reefDevicesKindReefBeat => 'Red Sea ReefBeat';
+  String get reefDevicesKindReefBeat => 'Red Sea';
 
   @override
   String get reefDevicesKindApex => 'Neptune Apex';
 
   @override
-  String get reefDevicesKindHanna => 'Hanna-Checker';
+  String get reefDevicesKindHanna => 'Hanna';
 
   @override
   String get reefDevicesBluetooth => 'Bluetooth';