+++ lib/l10n/app_localizations_it.dart
@@ -3806,13 +3806,13 @@ class AppLocalizationsIt extends AppLocalizations {
   String get reefDevicesKindReefFactory => 'ReefFactory';
 
   @override
-  String get reefDevicesKindReefBeat => 'Red Sea ReefBeat';
+  String get reefDevicesKindReefBeat => 'Red Sea';
 
   @override
   String get reefDevicesKindApex => 'Neptune Apex';
 
   @override
-  String get reefDevicesKindHanna => 'Checker Hanna';
+  String get reefDevicesKindHanna => 'Hanna';
 
   @override
   String get reefDevicesBluetooth => 'Bluetooth';