+++ lib/l10n/app_localizations_pl.dart
@@ -3866,13 +3866,13 @@ class AppLocalizationsPl extends AppLocalizations {
   String get reefDevicesKindReefFactory => 'ReefFactory';
 
   @override
-  String get reefDevicesKindReefBeat => 'Red Sea ReefBeat';
+  String get reefDevicesKindReefBeat => 'Red Sea';
 
   @override
   String get reefDevicesKindApex => 'Neptune Apex';
 
   @override
-  String get reefDevicesKindHanna => 'Checker Hanna';
+  String get reefDevicesKindHanna => 'Hanna';
 
   @override
   String get reefDevicesBluetooth => 'Bluetooth';